Implemented csf URL param for GET registrations
