+++ src/lib/orionld/notifications/subCacheAlterationMatch.cpp
@@ -31,7 +31,6 @@ extern "C"
 }
 
 #include "logMsg/logMsg.h"                                     // LM_*
-#include "logMsg/traceLevels.h"                                // LmtWatchedAttributes
 
 #include "common/sem.h"                                        // cacheSemTake, cacheSemGive
 #include "cache/subCache.h"                                    // CachedSubscription, subCacheMatch, tenantMatch
@@ -40,11 +39,16 @@ extern "C"
 #include "orionld/types/OrionldAlteration.h"                   // OrionldAlteration, OrionldAlterationMatch, orionldAlterationType
 #include "orionld/common/orionldState.h"                       // orionldState
 #include "orionld/common/dotForEq.h"                           // dotForEq
-#include "orionld/common/pathComponentsSplit.h"                // pathComponentsSplit
-#include "orionld/common/eqForDot.h"                           // eqForDot
 #include "orionld/common/dateTime.h"                           // dateTimeFromString
+#include "orionld/kjTree/kjNavigate.h"                         // kjNavigate2
 #include "orionld/q/qBuild.h"                                  // qBuild
 #include "orionld/q/qPresent.h"                                // qPresent
+#include "orionld/q/qRangeCompare.h"                           // qRangeCompare
+#include "orionld/q/qMatchCompare.h"                           // qMatchCompare
+#include "orionld/q/qCommaListCompare.h"                       // qCommaListCompare
+#include "orionld/q/qEqCompare.h"                              // qEqCompare
+#include "orionld/q/qGtCompare.h"                              // qGtCompare
+#include "orionld/q/qLtCompare.h"                              // qLtCompare
 #include "orionld/notifications/subCacheAlterationMatch.h"     // Own interface
 
 
@@ -262,7 +266,7 @@ static OrionldAlterationMatch* matchToMatchList
 //
 // falseUpdate -
 //
-bool falseUpdate(KjNode* attrP, KjNode* dbAttrsP)
+static bool falseUpdate(KjNode* attrP, KjNode* dbAttrsP)
 {
   char eqAttrName[512];
 
@@ -441,599 +445,9 @@ static OrionldAlterationMatch* attributeMatch(OrionldAlterationMatch* matchList,
 
 // -----------------------------------------------------------------------------
 //
-// dotCount - count number of dots in a string (no of components in a path)
+// qMatch -
 //
-static int dotCount(char* s)
-{
-  int dots = 0;
-
-  while (*s != 0)
-  {
-    if (*s == '.')
-      ++dots;
-    ++s;
-  }
-
-  return dots;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// kjNavigate - true Kj-Tree navigation
-//
-static KjNode* kjNavigate(KjNode* treeP, char** compV)
-{
-  KjNode* hitP = kjLookup(treeP, compV[0]);
-
-  if (hitP == NULL)
-    return NULL;
-
-  if (compV[1] == NULL)
-    return hitP;
-
-  return kjNavigate(hitP, &compV[1]);
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// kjNavigate2 - prepared for db-model, but also OK without
-//
-static KjNode* kjNavigate2(KjNode* treeP, char* path, bool* isTimestampP)
-{
-  int components = dotCount(path) + 1;
-  if (components > 20)
-    LM_X(1, ("The current implementation of Orion-LD can only handle 20 levels of tree navigation"));
-
-  char* compV[20];
-
-  // pathComponentsSplit destroys the path, I need to work on a copy
-  char* pathCopy = kaStrdup(&orionldState.kalloc, path);
-  components = pathComponentsSplit(pathCopy, compV);
-
-  //
-  // - the first component is always the longName of the ATTRIBUTE
-  // - the second is either "value", "object", "languageMap", or the longName of the SUB-ATTRIBUTE
-  //
-  // 'attrs' and 'md' don't exist in an API entity and must be nulled out here.
-  //
-
-  //
-  // Is it a timestamp?   (if so, an ISO8601 string must be turned into a float/integer to be compared
-  //
-  char* lastComponent = compV[components - 1];
-  if ((strcmp(lastComponent, "observedAt") == 0) || (strcmp(lastComponent, "modifiedAt") == 0) || (strcmp(lastComponent, "createdAt") == 0))
-    *isTimestampP = true;
-  else
-    *isTimestampP = false;
-
-  compV[components] = NULL;
-
-  KjNode* result = kjNavigate(treeP, compV);
-  if (result != NULL)
-    return result;
-
-  //
-  // Nothing found
-  //
-  //   What if it's due to '.' vs '=' ...
-  //   Yes, I know, this is messy, some order is needed
-  //
-  // FIXME: Fix this!
-  //
-  eqForDot(compV[0]);  // As it IS an Attribute
-  eqForDot(compV[1]);  // As it MIGHT be a Sub-Attribute (and if not, it has no '=')
-
-  result = kjNavigate(treeP, compV);
-  if (result != NULL)
-    return result;
-
-  //
-  // Could be a Relationship ...
-  // Perhaps I should "bake in" the value|object|languageMap inside kjNavigate ...
-  //
-  if ((components == 2) && (strcmp(compV[1], "value") == 0))
-  {
-    compV[1] = (char*) "object";
-    result = kjNavigate(treeP, compV);
-    if (result != NULL)
-      return result;
-
-    // FIXME: Here I put the '=' back ... Even messier now :(
-    dotForEq(compV[0]);
-    result = kjNavigate(treeP, compV);
-    if (result != NULL)
-      return result;
-  }
-
-  return NULL;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// qRangeCompare -
-//
-static bool qRangeCompare(KjNode* lhsNode, QNode* rhs, bool isTimestamp)
-{
-  QNode* low  = rhs->value.children;
-
-  if (low == NULL)
-    return false;
-
-  QNode* high = low->next;
-
-  if (high == NULL)
-    return false;
-
-  if (isTimestamp)
-  {
-    char errorString[256];
-
-    double lhsTimestamp  = dateTimeFromString(lhsNode->value.s, errorString, sizeof(errorString));
-    double lowTimestamp  = (low->type == QNodeFloatValue)? low->value.f :  dateTimeFromString(low->value.s, errorString, sizeof(errorString));
-    double highTimestamp = dateTimeFromString(high->value.s, errorString, sizeof(errorString));
-
-    if ((lhsTimestamp < 0) || (lowTimestamp < 0) || (highTimestamp < 0))
-      LM_RE(false, ("Invalid ISO8601 timestamp: %s", errorString));
-
-    if ((lhsTimestamp >= lowTimestamp) && (lhsTimestamp <= highTimestamp))
-      return true;
-
-    return false;
-  }
-
-  if (lhsNode->type == KjInt)
-  {
-    if (low->type == QNodeIntegerValue)
-    {
-      if (lhsNode->value.i < low->value.i)
-        return false;
-    }
-    else if (low->type == QNodeFloatValue)
-    {
-      if ((double) lhsNode->value.i < low->value.f)
-        return false;
-    }
-    else
-      return false;  // type mismatch
-
-    if (high->type == QNodeIntegerValue)
-    {
-      if (lhsNode->value.i > high->value.i)
-        return false;
-    }
-    else if (high->type == QNodeFloatValue)
-    {
-      if ((double) lhsNode->value.i > high->value.f)
-        return false;
-    }
-    else
-      return false;  // type mismatch
-
-    return true;
-  }
-
-  if (lhsNode->type == KjFloat)
-  {
-    if (low->type == QNodeFloatValue)
-    {
-      if (lhsNode->value.f < low->value.f)
-        return false;
-    }
-    else if (low->type == QNodeIntegerValue)
-    {
-      if (lhsNode->value.f < (double) low->value.i)
-        return false;
-    }
-    else
-      return false;  // type mismatch
-
-    if (high->type == QNodeFloatValue)
-    {
-      if (lhsNode->value.f > high->value.f)
-        return false;
-    }
-    else if (high->type == QNodeIntegerValue)
-    {
-      if (lhsNode->value.f > (double) high->value.i)
-        return false;
-    }
-    else
-      return false;  // type mismatch
-
-    return true;
-  }
-
-  if (lhsNode->type == KjString)
-  {
-    if (low->type == QNodeStringValue)
-    {
-      if (strcmp(lhsNode->value.s, low->value.s)  < 0)
-        return false;
-      if (strcmp(lhsNode->value.s, high->value.s) > 0)
-        return false;
-
-      return true;
-    }
-
-    return false;  // type mismatch
-  }
-
-  // Timestamp!
-
-  return false;  // RANGES operate only on Numbers, Strings and Timestamps
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// intComparison
-//
-bool intComparison(long long lhs, QNode* rhs)
-{
-  if      (rhs->type == QNodeIntegerValue)  return lhs == rhs->value.i;
-  else if (rhs->type == QNodeFloatValue)    return ((double) lhs) == rhs->value.f;
-
-  return false;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// floatComparison
-//
-bool floatComparison(KjNode* lhsP, QNode* rhs, bool isTimestamp)
-{
-  if (isTimestamp == false)
-  {
-    double lhs = lhsP->value.f;
-
-    if      (rhs->type == QNodeFloatValue)    return lhs == rhs->value.f;  // precision for float comparison??
-    else if (rhs->type == QNodeIntegerValue)  return lhs == (double) rhs->value.i;
-    else                                      return false;
-  }
-
-  if (rhs->type == QNodeStringValue)
-  {
-    char   errorString[256];
-    double rhsTimestamp = dateTimeFromString(rhs->value.s, errorString, sizeof(errorString));
-
-    if (lhsP->value.f == rhsTimestamp)
-      return true;
-  }
-
-  return false;
-}
-
-
-// -----------------------------------------------------------------------------
-//
-// boolComparison
-//
-bool boolComparison(bool lhs, QNode* rhs)
-{
-  if ((lhs == true) && (rhs->type == QNodeTrueValue))
-    return true;
-
-  if ((lhs == false) && (rhs->type == QNodeFalseValue))
-    return true;
-
-  return false;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// stringComparison
-//
-bool stringComparison(KjNode* lhsP, QNode* rhs, bool isTimestamp)
-{
-  if (rhs->type != QNodeStringValue)
-    return false;
-
-  if (isTimestamp)  // Then LHS is a Float?
-  {
-    char   errorString[256];
-    double rhsTimestamp = dateTimeFromString(rhs->value.s, errorString, sizeof(errorString));
-
-    return (lhsP->value.f == rhsTimestamp);
-  }
-
-  if (rhs->type != QNodeStringValue)
-    return false;
-
-  return (strcmp(lhsP->value.s, rhs->value.s) == 0);
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// qCommaListCompare -
-//
-bool qCommaListCompare(KjNode* lhsNode, QNode* rhs, bool isTimestamp)
-{
-  if (isTimestamp == true)  // The first in the list id a FLOAT, the rest need to be converted to float
-  {
-    double timestamp;
-    char   errorString[256];
-
-    if (lhsNode->type == KjFloat)
-      timestamp = lhsNode->value.f;
-    else if (lhsNode->type == KjString)
-      timestamp = dateTimeFromString(lhsNode->value.s, errorString, sizeof(errorString));
-    else
-      return false;
-
-    QNode* child1 = rhs->value.children;
-
-    if (child1->type != QNodeFloatValue)
-      LM_E(("CLIST: Internal Error (LHS is a timestamp but the first in the RHS list is not a FLOAT ..."));
-    else if (child1->value.f == timestamp)
-      return true;
-
-    // Compare the rest of the children in the comma list (converting them to FLOAT)
-    for (QNode* rhP = child1->next; rhP != NULL; rhP = rhP->next)
-    {
-      if (rhP->type == QNodeFloatValue)
-      {
-        if (rhP->value.f == timestamp)
-          return true;
-      }
-      else if (rhP->type == QNodeStringValue)
-      {
-        char   errorString[256];
-        double rhsTimestamp = dateTimeFromString(rhP->value.s, errorString, sizeof(errorString));
-
-        if (rhsTimestamp == timestamp)
-          return true;
-      }
-    }
-
-    return false;
-  }
-
-  for (QNode* rhP = rhs->value.children; rhP != NULL; rhP = rhP->next)
-  {
-    switch (lhsNode->type)
-    {
-    case KjInt:     if (intComparison(lhsNode->value.i,    rhP)              == true) return true; break;
-    case KjFloat:   if (floatComparison(lhsNode,           rhP, isTimestamp) == true) return true; break;  // isTimestamp - then LHS is turned into a Float ... Right?
-    case KjString:  if (stringComparison(lhsNode,          rhP, isTimestamp) == true) return true; break;  //               Can't be both - need to test empirically
-    case KjBoolean: if (boolComparison(lhsNode->value.b,   rhP)              == true) return true; break;
-    default:
-      break;
-    }
-  }
-
-  return false;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// qEqCompare -
-//
-bool qEqCompare(KjNode* lhsNode, QNode* rhs, bool isTimestamp)
-{
-  //
-  // Might be a timestamp ... (observedAt, modifiedAt, or createdAt)
-  //
-  if (isTimestamp == true)
-  {
-    // lhsNode must be a string, and a valid ISO8601 at that
-    if (lhsNode->type != KjString)
-      return false;
-
-    char   errorString[256];
-    double timestamp = dateTimeFromString(lhsNode->value.s, errorString, sizeof(errorString));
-
-    if (rhs->type == QNodeIntegerValue)
-    {
-      long long ts = (long long) timestamp;
-      if (rhs->value.i == ts)
-        return true;
-
-      return false;
-    }
-    else if (rhs->type == QNodeFloatValue)
-    {
-      if (rhs->value.f == timestamp)
-        return true;
-
-      return false;
-    }
-  }
-
-  if (lhsNode->type == KjInt)
-  {
-    if      (rhs->type == QNodeIntegerValue) return (lhsNode->value.i == rhs->value.i);
-    else if (rhs->type == QNodeFloatValue)   return (lhsNode->value.i == rhs->value.f);
-  }
-  else if (lhsNode->type == KjFloat)
-  {
-    double margin = 0.000001;  // What margin should I use?
-
-    if (rhs->type == QNodeIntegerValue)
-    {
-      if ((lhsNode->value.f - margin < rhs->value.i) && (lhsNode->value.f + margin > rhs->value.i))
-        return true;
-    }
-    else if (rhs->type == QNodeFloatValue)
-    {
-      if ((lhsNode->value.f - margin < rhs->value.f) && (lhsNode->value.f + margin > rhs->value.f))
-        return true;
-    }
-  }
-  else if (lhsNode->type == KjString)
-  {
-    if (rhs->type == QNodeStringValue)
-      return (strcmp(lhsNode->value.s, rhs->value.s) == 0);
-  }
-  else if (lhsNode->type == KjBoolean)
-  {
-    if (rhs->type == QNodeTrueValue)       return (lhsNode->value.b == true);
-    if (rhs->type == QNodeFalseValue)      return (lhsNode->value.b == false);
-  }
-
-  return false;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// qGtCompare -
-//
-bool qGtCompare(KjNode* lhsNode, QNode* rhs, bool isTimestamp)
-{
-  //
-  // Might be a timestamp ... (observedAt, modifiedAt, or createdAt)
-  //
-  if (isTimestamp == true)
-  {
-    // lhsNode must be a string, and a valid ISO8601 at that
-    if (lhsNode->type != KjString)
-      return false;
-
-    char   errorString[256];
-    double timestamp = dateTimeFromString(lhsNode->value.s, errorString, sizeof(errorString));
-
-    if (rhs->type == QNodeIntegerValue)
-    {
-      long long ts = (long long) timestamp;
-      if (rhs->value.i < ts)
-        return true;
-
-      return false;
-    }
-    else if (rhs->type == QNodeFloatValue)
-    {
-      if (rhs->value.f < timestamp)
-        return true;
-
-      return false;
-    }
-  }
-
-  if (lhsNode->type == KjInt)
-  {
-    if      (rhs->type == QNodeIntegerValue) return (lhsNode->value.i > rhs->value.i);
-    else if (rhs->type == QNodeFloatValue)   return (lhsNode->value.i > rhs->value.f);
-  }
-  else if (lhsNode->type == KjFloat)
-  {
-    if      (rhs->type == QNodeIntegerValue) return (lhsNode->value.f > rhs->value.i);
-    else if (rhs->type == QNodeFloatValue)   return (lhsNode->value.f > rhs->value.f);
-  }
-  else if (lhsNode->type == KjString)
-  {
-    if (rhs->type == QNodeStringValue)
-      return (strcmp(lhsNode->value.s, rhs->value.s) > 0);  // "> 0" means first arg > second arg to strcmp
-  }
-  else if (lhsNode->type == KjBoolean)  // true > false ... ?
-  {
-    if (rhs->type == QNodeTrueValue)       return false;
-    if (rhs->type == QNodeFalseValue)      return (lhsNode->value.b == true);
-  }
-
-  return false;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// qLtCompare -
-//
-bool qLtCompare(KjNode* lhsNode, QNode* rhs, bool isTimestamp)
-{
-  //
-  // Might be a timestamp ... (observedAt, modifiedAt, or createdAt)
-  //
-  if (isTimestamp == true)
-  {
-    // lhsNode must be a string, and a valid ISO8601 at that
-    if (lhsNode->type != KjString)
-      return false;
-
-    char   errorString[256];
-    double timestamp = dateTimeFromString(lhsNode->value.s, errorString, sizeof(errorString));
-
-    if (rhs->type == QNodeIntegerValue)
-    {
-      long long ts = (long long) timestamp;
-      if (rhs->value.i > ts)
-        return true;
-
-      return false;
-    }
-    else if (rhs->type == QNodeFloatValue)
-    {
-      if (rhs->value.f > timestamp)
-        return true;
-
-      return false;
-    }
-  }
-
-  bool r = false;
-  if (lhsNode->type == KjInt)
-  {
-    if      (rhs->type == QNodeIntegerValue) r = (lhsNode->value.i < rhs->value.i);
-    else if (rhs->type == QNodeFloatValue)   r = (lhsNode->value.i < rhs->value.f);
-  }
-  else if (lhsNode->type == KjFloat)
-  {
-    if      (rhs->type == QNodeIntegerValue) r = (lhsNode->value.f < rhs->value.i);
-    else if (rhs->type == QNodeFloatValue)   r = (lhsNode->value.f < rhs->value.f);
-  }
-  else if (lhsNode->type == KjString)
-  {
-    if (rhs->type == QNodeStringValue)
-      r = (strcmp(lhsNode->value.s, rhs->value.s) < 0);  // "< 0" means first arg < second arg to strcmp
-  }
-  else if (lhsNode->type == KjBoolean)  // true > false ... ?
-  {
-    if (rhs->type == QNodeTrueValue)       r = false;
-    if (rhs->type == QNodeFalseValue)      r = (lhsNode->value.b == true);
-  }
-
-  return r;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// qMatchCompare -
-//
-bool qMatchCompare(KjNode* lhsNode, QNode* rhs)
-{
-  //
-  // Create the REGEX in subCache - QNode might need a new QNodeValue
-  // Here: Use regexec
-  //
-  LM_W(("Not Implemented"));
-  return false;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// qMatch - move to orionld/q/qMatch.h/cpp
-//
-bool qMatch(QNode* qP, OrionldAlteration* altP)
+static bool qMatch(QNode* qP, OrionldAlteration* altP)
 {
   if (qP->type == QNodeOr)
   {
@@ -1060,7 +474,7 @@ bool qMatch(QNode* qP, OrionldAlteration* altP)
   else
   {
     QNode* lhs = qP->value.children;        // variable-path
-    QNode* rhs = qP->value.children->next;  // constant (MULL for QNodeExists & QNodeNotExists
+    QNode* rhs = qP->value.children->next;  // constant (NULL for QNodeExists & QNodeNotExists)
 
     //
     // Not OR nor AND => LHS is an  attribute from the entity
