+++ src/lib/orionld/q/CMakeLists.txt
@@ -36,6 +36,12 @@ SET (SOURCES
     qLexListRender.cpp
     qClone.cpp
     qVariableFix.cpp
+    qMatchCompare.cpp
+    qRangeCompare.cpp
+    qCommaListCompare.cpp
+    qEqCompare.cpp
+    qGtCompare.cpp
+    qLtCompare.cpp
 )
 
 # Include directories
