+++ test/functionalTest/cases/0000_ngsild/ngsild_new_registration_and_context_02.test
@@ -248,7 +248,7 @@ MongoDB server version: REGEX(.*)
 	},
 	"expiration" : 1861869601.234,
 	"properties" : {
-		"https://uri.etsi.org/ngsi-ld/default-context/P1" : 1
+		"https://uri=etsi=org/ngsi-ld/default-context/P1" : 1
 	},
 	"status" : "active"
 }
