+++ test/functionalTest/cases/0000_ngsild/ngsild_registration_and_context_01.test
@@ -247,7 +247,7 @@ MongoDB server version: REGEX(.*)
 		]
 	},
 	"properties" : {
-		"https://uri.etsi.org/ngsi-ld/default-context/P1" : 1
+		"https://uri=etsi=org/ngsi-ld/default-context/P1" : 1
 	}
 }
 bye
