+++ test/functionalTest/cases/0000_ngsild/ngsild_registration_and_context_02.test
@@ -251,7 +251,7 @@ MongoDB server version: REGEX(.*)
 		]
 	},
 	"properties" : {
-		"https://uri.etsi.org/ngsi-ld/default-context/P1" : 1
+		"https://uri=etsi=org/ngsi-ld/default-context/P1" : 1
 	}
 }
 bye
